// app/src/api/streak_board/StreakBoardClient.ts
export interface Habit {
  id: string;
  author: string;
  title: string;
  current_streak: number;
  longest_streak: number;
  last_check_in_date: string;
  created_at: number;
}

export interface CheckIn {
  id: string;
  author: string;
  habit_id: string;
  date: string;
  timestamp: number;
}

export interface Cheer {
  id: string;
  author: string;
  habit_id: string;
  message: string;
  created_at: number;
}